botPython/requirements.txt
# frameworks da BotCity
botcity-framework-web>=0.8.0,<1.0
botcity-maestro-sdk>=0.3.3,<1.0

# Plugins
botcity-email-plugin
botcity-googlesheets-plugin

python-dotenv